requirements.txt
flask
redis
kafka-python
faker
